Update to CSS and HTML
I focused on CSS this time to make the UI feel nice, which scales well on phone too. Checked for WAVE mistakes and corrected it (low contrast, missing labels). Temporarily removed canvas element. Might implement it later again with three.js, but commented out the code for now. Separated the initial choice logic (HTML) into a template and injected it from JS file instead. This is to make it look cleaner, since it will be removed after the choices are made anyway.